feat: removing providers and introducing useBlob/search/db helpers
