+++ packages/gensx-storage/src/index.ts
@@ -1,19 +1,14 @@
 // Export the blob storage types and interfaces
 export * from "./blob/types.js";
-export * from "./blob/context.js";
-export { BlobProvider } from "./blob/provider.js";
+export { useBlob } from "./blob/useBlob.js";
 export { BlobClient } from "./blob/blobClient.js";
 
+// Export the database types and interfaces
 export * from "./database/types.js";
-export * from "./database/context.js";
-export { DatabaseProvider } from "./database/provider.js";
+export { useDatabase } from "./database/useDatabase.js";
 export { DatabaseClient } from "./database/databaseClient.js";
-export * from "./search/types.js";
-export * from "./search/context.js";
-export { SearchProvider } from "./search/provider.js";
-export { SearchClient } from "./search/searchClient.js";
 
-// Re-export for convenience
-export { useBlob, useBlobStorage } from "./blob/context.js";
-export { useDatabase, useDatabaseStorage } from "./database/context.js";
-export { useSearch, useSearchStorage } from "./search/context.js";
+// Export the search types and interfaces
+export * from "./search/types.js";
+export { useSearch } from "./search/useSearch.js";
+export { SearchClient } from "./search/searchClient.js";
